project: remove qtribbon
*** NOTES ***
QTitanRibbon is a commercial component, this version (3.1.1) is from the Internet, published here for reference only,
prohibited in the absence of compliance with its licensing authority, violators at their own risk.
In order to narrow the diffusion range, it is removed.

QTitanRibbon, of historical records also follow the above statement.